allowed publicationStatus field on post creation and update, also added role check

+++ backend/controllers/postController.ts
@@ -1,5 +1,6 @@
 import type { NextFunction, Request, Response } from "express";
 import postsRepository from "../repositories/postsRepository";
+import { matchedData } from "express-validator";
 
 async function getPosts(
 	req: Request,
@@ -27,8 +28,8 @@ async function createPost(
 			return;
 		}
 		const userId: string = req.user?.id as string;
-		const { title, body } = req.body;
-		await postsRepository.createPost(title, body, userId);
+		const { title, body, publicationStatus } = matchedData(req);
+		await postsRepository.createPost(title, body, publicationStatus, userId);
 		res.status(201).json("POST CREATED");
 		return;
 	} catch (err) {
@@ -42,7 +43,7 @@ async function getPost(
 	next: NextFunction,
 ): Promise<void> {
 	try {
-		const postId: string = req.params.postId as string;
+		const { postId } = matchedData(req);
 		const post = await postsRepository.getPost(postId);
 		if (!post) {
 			res.status(404).json("The specified post doesn't exist");
@@ -60,9 +61,21 @@ async function updatePost(
 	next: NextFunction,
 ): Promise<void> {
 	try {
-		const postId: string = req.params.postId as string;
+		const role: string = req.user?.role as string;
+		if (role !== "ADMIN" && role !== "AUTHOR") {
+			res
+				.status(403)
+				.json("Cannot update post if you're not an author or an admin");
+			return;
+		}
+		const data = matchedData(req);
 		const userId: string = req.user?.id as string;
-		const newData: Object = req.body;
+		const postId: string = data.postId;
+		const newData = {
+			title: data["title"],
+			body: data["body"],
+			publicationStatus: data["publicationStatus"],
+		};
 		await postsRepository.updatePost(postId, newData, userId);
 		res.status(201).json("POST UPDATED");
 		return;
@@ -77,7 +90,14 @@ async function deletePost(
 	next: NextFunction,
 ): Promise<void> {
 	try {
-		const postId: string = req.params.postId as string;
+		const role: string = req.user?.role as string;
+		if (role !== "ADMIN" && role !== "AUTHOR") {
+			res
+				.status(403)
+				.json("Cannot delete post if you're not an author or an admin");
+			return;
+		}
+		const { postId } = matchedData(req);
 		const userId: string = req.user?.id as string;
 		await postsRepository.deletePost(postId, userId);
 		res.json("POST DELETED");
